feat: consentimiento WhatsApp real, tracking de costos por categoría y transcripción/audio de llamadas
Rehace el flujo de consentimiento previo a la llamada de profiling (esperar
aceptación/rechazo/timeout de WhatsApp en vez de llamar de inmediato), corrige
varios bugs de máquina de estados y CostLog que dejaban candidatos atascados o
disparaban mensajes duplicados, agrega watchdog de timeouts de WhatsApp al
beat schedule (nunca corría), fusiona el análisis y enriquecimiento de JD en
un solo flujo consciente del contexto del proceso, agrega desglose de costos
por categoría (voz/twilio/whatsapp/llm) y por candidato, persiste la
transcripción completa (turn-by-turn) de ElevenLabs y expone un endpoint para
streamear el audio de la llamada bajo demanda, y agrega instrucciones de
idioma español a los prompts de match/JD/evaluación de profiling.

Incluye además un endpoint DELETE /users/{id} (admin, con audit log) que
estaba en el working tree sin commitear.

+++ src/config.py
@@ -57,7 +57,7 @@ class Settings(BaseSettings):
 
     # Business rules
     max_concurrent_calls: int = 4
-    whatsapp_consent_timeout_hours: int = 24
+    whatsapp_consent_timeout_hours: float = 24
     cv_batch_limit: int = 50
     max_cv_file_size_mb: int = 10
     profiling_delay_seconds: int = 86400  # Default 24h
